requirements.txt
flask
gunicorn
scikit-learn
joblib
openai>=1.0.0
python-dotenv
